feat: redesign navigation with unified collapsible sidebar
- Replace tab-based navigation with unified sidebar (similar to ChatGPT/Grok)
- Add collapsible sidebar with icons-only view and bottom toggle button
- Move Chat/Tasks navigation from top menu bar to left sidebar
- Create MainLayout component that handles both chat and tasks views
- Create UnifiedSidebar component with improved spacing and layout
- Remove extra padding and improve content area utilization
- Remove superseded LeftSidebar, Conversations, and TaskManager components
- Add leftSidebarCollapsed$ state management to sidebar store

This provides a cleaner, more intuitive navigation experience with better
mobile responsiveness and follows modern UI patterns.
